Timing info added for
- map loading
- input loading
- event/stream creation
- event reset

The event reset time ("Reset") is included in both the "Event" and the "Short" time. 

**Timing for 100 events**

This PR (0b91428f653d32a2811b7a63bb4e28106552fc62):
```
Total Timing Summary
Average time for map loading = 3163.12 ms
Average time for input loading = 5536.64 ms
Average time for stream creation = 0.164747 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg      4.6      3.2      1.1      3.4      4.0      1.2      1.9      1.5      4.0      0.1      24.9      19.1+/-  3.5      29.1   explicit_cache[s=1]
   avg      7.6      4.6      1.7      6.6      5.6      1.5      3.7      2.4      6.5      0.1      40.4      31.4+/-  6.0      22.4   explicit_cache[s=2]
   avg     14.7      7.4      3.0     13.0     10.1      2.0      8.0      5.1     12.6      0.2      76.1      59.4+/- 10.8      20.4   explicit_cache[s=4]
   avg     24.1     10.1      4.8     18.5     15.1      2.7     12.5      7.3     18.0      0.4     113.7      86.9+/- 13.3      20.3   explicit_cache[s=6]
   avg     34.1     12.5      6.5     22.8     20.9      3.1     16.5      9.5     23.3      0.5     149.7     112.5+/- 19.2      20.2   explicit_cache[s=8]

```

Master (6e6e2d0612a9c67223b72da25fb5bf5ff4b6c848):
```
Total Timing Summary
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Event      Short           Rate
   avg      4.6      3.1      1.0      3.3      3.8      1.2      1.8      1.5      3.6      23.9      18.1+/-  3.4      28.1   explicit_cache[s=1]
   avg      7.7      4.4      1.7      6.4      5.5      1.5      3.5      2.4      6.5      39.7      30.5+/-  5.5      22.1   explicit_cache[s=2]
   avg     14.5      7.9      3.1     13.5      9.8      2.0      8.3      5.0     12.8      76.8      60.3+/- 10.0      20.5   explicit_cache[s=4]
   avg     22.1     10.7      4.7     18.6     15.8      2.7     13.6      7.4     18.1     113.5      88.8+/- 12.9      20.4   explicit_cache[s=6]
   avg     32.1     13.0      6.6     23.8     20.9      3.1     16.8      9.2     23.1     148.5     113.4+/- 21.0      20.3   explicit_cache[s=8]

```