@GNiendorf 2023-07-13
Did you ever figure out what the discrepancy between the event time and rate time was caused by btw? I think you may have mentioned it in the last meeting but I forget now.