@GNiendorf 2023-07-13
> > Did you ever figure out what the discrepancy between the event time and rate time was caused by btw?
> 
> No, I don't have an answer to that. However, there was a question and maybe you are the best person to try and answer it: "Are the `cms::cuda::free_host()` calls synchronous?"

I'm not sure to be honest. I haven't looked into the CUDA caching allocator that much, and in alpaka everything is freed when it goes out of scope so I haven't paid much attention to it.