@GNiendorf 2023-07-15
As a last comment, is there any way that you could separate the map loading time from the initModules function time? I'm curious how much of the loadMaps time is actually loading the maps vs this sort of hidden function that was relevant in your last commit. I'm not sure if you have anything in mind, but one solution would be changing loadMaps to return the centroid variable so that you can run initModules from sdl.cc instead of having this function call hidden away like it is now. Thank you!: https://github.com/SegmentLinking/TrackLooper/blob/6e6e2d0612a9c67223b72da25fb5bf5ff4b6c848/code/core/trkCore.cc#L56