@VourMa 2023-07-15
>  is there any way that you could separate the map loading time from the initModules function time?

I wouldn't start refactorizing the code just to add more timing measurements, so I would prefer to keep this as it is in the standalone.

That said, I am not saying that the question you ask is not interesting. There is another way to get a feeling of the timing of the different calls. In the CMSSW implementation, I have separated the loading of the maps ([`SDL::LST::loadMaps()`](https://evourlio.web.cern.ch/evourlio/cgi-bin/igprof-navigator/ig_LSTinCMSSW_6e6e2d0_phi3_100Events/29)) and the initiation of the modules ([`SDL::initModules`](https://evourlio.web.cern.ch/evourlio/cgi-bin/igprof-navigator/ig_LSTinCMSSW_6e6e2d0_phi3_100Events/720)). In the [igprof profiling I did recently for 1000 events](https://evourlio.web.cern.ch/evourlio/cgi-bin/igprof-navigator/ig_LSTinCMSSW_6e6e2d0_phi3_100Events/28), one can see that `initModules` takes ~1/100 of the time that `loadMaps` takes.