@VourMa 2023-07-15
Timing after [6e2b1ba](https://github.com/SegmentLinking/TrackLooper/pull/305/commits/6e2b1baf442836cf1b857b4899777a49820d01b5)
```
Total Timing Summary
Average time for map loading = 3172.06 ms
Average time for input loading = 5500.08 ms
Average time for SDL::Event creation = 0.0395094 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg      4.5      3.1      1.0      3.5      4.0      1.2      1.8      1.5      3.7      0.1      24.4      18.7+/-  3.5      28.6   explicit_cache[s=1]
   avg      7.1      4.6      1.5      6.4      5.4      1.4      3.6      2.5      6.0      0.1      38.6      30.2+/-  5.5      21.5   explicit_cache[s=2]
   avg     13.9      7.3      3.0     12.7     10.0      2.0      8.3      4.9     11.9      0.2      74.3      58.4+/-  9.2      19.8   explicit_cache[s=4]
   avg     21.6     10.5      4.9     18.0     15.5      2.7     12.8      7.2     18.1      0.3     111.6      87.3+/- 13.2      20.2   explicit_cache[s=6]
   avg     32.6     13.2      6.4     23.5     20.9      3.5     17.5      9.7     22.3      0.5     150.0     113.9+/- 16.1      19.9   explicit_cache[s=8]
```