@GNiendorf 2023-07-15
> > is there any way that you could separate the map loading time from the initModules function time?
> 
> I wouldn't start refactorizing the code just to add more timing measurements, so I would prefer to keep this as it is in the standalone.

My main concern was that it doesn't make sense to me to have initmodules hidden within the "loadmaps" function (and could be misleading for the loadmaps average time), but my first point might be outside the scope of this PR. If initmodules takes 1/100th the time of the map loading then I'm okay with leaving things as is for this PR. When you push the change related to the SDL::Event time I'll merge this in.